+++ src/task/watchLink/watch-link.service.ts
@@ -1,6 +1,6 @@
 import { TaskConfig } from '@/config';
 import { getUser } from '@/net/user-info.request';
-import { biliDmWs, bindMessageForRedPacket } from '@/service/ws.service';
+import { biliDmWs } from '@/service/ws.service';
 import { logger } from '@/utils';
 import { apiDelaySync } from '@/utils/effect';
 import { request } from '@/utils/request';
@@ -101,10 +101,10 @@ async function getUserInfo(uid: number | string) {
 async function createWatchDmWs({ room_id }, wsTime: number) {
   const ws = await biliDmWs(room_id, (wsTime + 20) * 1000 * 60);
   if (!ws) return;
-  bindMessageForRedPacket(ws, room_id, async body => {
-    if (body.cmd !== 'STOP_LIVE_ROOM_LIST') {
-      console.log(room_id, JSON.stringify(body));
-    }
-  });
+  // bindMessageForRedPacket(ws, room_id, async body => {
+  //   if (body.cmd !== 'STOP_LIVE_ROOM_LIST') {
+  //     console.log(room_id, JSON.stringify(body));
+  //   }
+  // });
   return ws;
 }
